Fix MaaS testing for AIO gating
Currently the MaaS testing does not form part of a gate, this has caused
the testing code to diverge from the deployed monitoring. This commit
addresses a number of identified issues and makes AIO testing the
default action.

The generate-definitions.py script has been updated to allow specific
attributes to be templated. This is required to allow a password in the
RabbitMQ check to be replaced following an update to how that check is
defined.

Rackspace Cloud servers now automatically get a ping check created. The
test-maas.yml playbook will delete any existing checks from the
entities.

MaaS remote checks have a new attribute called collectors. The
collectors attribute will be removed because it does not need verifying.
The attribute confd_hash is also removed, it relates to the conf files
that are used by the agent to create the required checks/alarms.