Add Pending-SHA-update column to Waffle board
In general an upstream fix is not immediately consumable by RPCO, OSA
has processes to periodically update its SHAs to pull in newer code from
associated projects. This lag between a fix appearing in a code base and
being available to RPCO has the potential to block new work. This commit
adds documentation for a new column, Pending-SHA-update, that can be
used as a buffer so that this work is appropriately tracked but does not
block new WIP.