Remove priority labels and add severity labels
Priority is a measure of the importance of fixing an issue, severity
describes the impact an issue has on the functionality of the software.

Currently a priority label is assigned in addition to the prioritisation
that occurs when the issues become approved. The priority label is
currently describing both priority and severity.

This commit removes the use of priority labels and adds severity labels.
This creates a separation between the two pieces of information and
leaves priority to be the domain of the Kanban board. An expedite label
is kept to allow for expedited issues to be easily tracked once they
have moved past the point of commitment.