Do some verification before starting upgrade tasks
We need to do some verification before starting upgrade tasks in
order to avoid errors on subsequent runs. First off we install a
pre flight check script that gathers version information and
returns codes based on the currently installed version. A pre-flight
task will execute this script and will set a flag to skip the upgrade
if the latest version is installed and exit if the reindexing version
is installed.

I have also created variables elasticsearch_version and
elasticsearch_reindex_version so that we can hard code the versions
of elasticsearch to be used. We were using hardcoded and assumed
versions in several places.

Thirdly I have added a retries to the curl command in order to
skip a transient error that we have occasionally seen.

Connects #1512