Removed fact check and ignore errors
Since the facts are not globally available there isn't a way to
change the runtime behavior once the play has started with
logging_upgrade set to true.  This removes the ignore errors so
an attempt to run the upgrade with either the latest version of
elasticsearch installed or the reindexing version will abort
the upgrade.