Change version placement and add force
The elasticsearch version was being applied to all packages in the
list so I moved it to specify the version in the list. I also created
a package list for the reindexing version of elasticsearch.

Added a force option so that the upgrade play can be run regardless
of the output of the pre-flight checks. This will be needed when a
long running reindexing process is started prior to the actual upgrade.